Refactor database layer with modular schema and store architecture
- Reorganize database structure with separate schema, store, and types directories
- Remove monolithic schema.ts and urlStore.ts in favor of modular approach
- Add comprehensive database schema organization with separate files
- Implement modular store pattern for better data access layer
- Update drizzle configuration for new database structure
- Enhance middleware with improved database integration
- Update routes to use new modular database stores
- Improve templates with better styling and functionality
- Update TypeScript configuration for better module resolution
- Add proper type definitions for database entities
- Improve code organization and maintainability
