+++ drizzle.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
   out: './drizzle',
-  schema: './src/db/schema.ts',
+  schema: './src/db/schema',
   dialect: 'sqlite',
   dbCredentials: {
     url: process.env.DB_FILE_NAME!,
