+++ src/db/schema/session.ts
@@ -1,0 +1,7 @@
+import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core';
+
+export const sessionTable = sqliteTable('session', {
+  id: int().primaryKey({ autoIncrement: true }),
+  code: text().notNull().unique(),
+  expires: int().notNull(),
+});
