+++ src/db/types/shortlink.ts
@@ -1,0 +1,3 @@
+import { shortlinkTable } from '../schema/shortlink.js';
+
+export type Shortlink = typeof shortlinkTable.$inferSelect;
