+++ src/middleware/no-auth.ts
@@ -1,4 +1,4 @@
-import sessionStore from '@/sessionStore.js';
+import sessionStore from '@/db/store/session.js';
 import { wantsHtml } from '@/util/html.js';
 import { SESSION_COOKIE } from '@/util/session.js';
 import type { Context, Next } from 'hono';
