+++ src/routes/admin.ts
@@ -1,9 +1,9 @@
+import sessionStore from '@/db/store/session.js';
+import urlStore from '@/db/store/shortlink.js';
 import { requireAuth } from '@/middleware/auth.js';
 import { requireNoAuth } from '@/middleware/no-auth.js';
-import sessionStore from '@/sessionStore.js';
 import DashboardPage from '@/templates/dashboard.js';
 import LoginPage from '@/templates/login.js';
-import urlStore from '@/urlStore.js';
 import { getCode } from '@/util/code.js';
 import { wantsHtml } from '@/util/html.js';
 import { SESSION_COOKIE } from '@/util/session.js';
