+++ tsconfig.json
@@ -14,7 +14,6 @@
     "paths": {
       "@/*": ["src/*"]
     }
-    
   },
   "exclude": ["node_modules"]
 }
